requirements.txt
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.7.0
requests>=2.31.0
beautifulsoup4>=4.12.2
newspaper3k>=0.2.8
lxml[html_clean]>=4.9.3
feedparser>=6.0.10
nltk>=3.8.1
tldextract>=5.1.2
Pillow>=10.0.0
python-dateutil>=2.9.0
certifi>=2023.11.17
chardet>=5.2.0
idna>=3.7
urllib3>=2.2.2